+++ src/webgl/mesh/type.ts
@@ -1,0 +1,2 @@
+
+export type colorType = number[] | number;
